+++ requirements.txt
@@ -49,3 +49,4 @@ spaCy==3.6.1
 justext==3.0.0
 jieba==0.42.1
 langdetect==1.0.9
+ray==2.6.3
